色々追加

+++ src/krapp/diary_organizer.py
@@ -10,7 +10,7 @@ class DiaryOrganizer:
         self.date_extractor = date_extractor
 
     def organize_diary(
-        self, input_md_file: str | Path, output_folder: str | Path
+        self, input_md_file: str | Path, output_folder: str | Path, remove: bool = False
     ) -> None:
         # Read the content of the input markdown file
         input_md_file = Path(input_md_file)
@@ -18,7 +18,9 @@ class DiaryOrganizer:
         content = input_md_file.read_text(encoding="utf-8")
 
         # Extract the date from the content
-        date_list = self.date_extractor.extract_dates(content)
+        date_list = self.date_extractor.extract_dates(
+            f"{input_md_file.stem}\n{content}"
+        )
         if len(date_list) == 0:
             raise ValueError("No valid date found in the markdown file.")
         date = date_list[0]
@@ -28,5 +30,13 @@ class DiaryOrganizer:
         month_folder = year_folder / f"{date.month:02d}"
         os.makedirs(month_folder, exist_ok=True)
 
+        # 同じ名前のファイルがあった場合は処理をスキップ
+        if month_folder.joinpath(input_md_file.name).exists():
+            print(
+                f"File {input_md_file.name} already exists in {month_folder}. Skipping."
+            )
+            return
         # Copy the markdown file to the appropriate folder
         shutil.copy(input_md_file, month_folder)
+        if remove:
+            os.remove(input_md_file)
